+++ commons-civil/src/main/java/uk/gov/hmcts/reform/hmc/model/unnotifiedhearings/PartiesNotifiedResponses.java
@@ -1,18 +1,21 @@
-package uk.gov.hmcts.reform.hmc.model.unnotifiedHearings;
+package uk.gov.hmcts.reform.hmc.model.unnotifiedhearings;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class UnNotifiedHearingResponse {
+public class PartiesNotifiedResponses {
 
-    private List<String> hearingIds;
+    private String hearingID;
 
-    private Long totalFound;
+    private List<PartiesNotifiedResponse> responses;
+
 }
+
