+++ commons-civil/src/test/java/uk/gov/hmcts/reform/hmc/service/HearingsServiceTest.java
@@ -20,9 +20,9 @@ import uk.gov.hmcts.reform.hmc.model.hearing.HearingGetResponse;
 import uk.gov.hmcts.reform.hmc.model.hearing.HearingRequestDetails;
 import uk.gov.hmcts.reform.hmc.model.hearing.HearingResponse;
 import uk.gov.hmcts.reform.hmc.model.hearing.PartyDetailsModel;
-import uk.gov.hmcts.reform.hmc.model.unnotifiedHearings.PartiesNotifiedResponse;
-import uk.gov.hmcts.reform.hmc.model.unnotifiedHearings.PartiesNotifiedResponses;
-import uk.gov.hmcts.reform.hmc.model.unnotifiedHearings.UnNotifiedHearingResponse;
+import uk.gov.hmcts.reform.hmc.model.unnotifiedhearings.PartiesNotifiedResponse;
+import uk.gov.hmcts.reform.hmc.model.unnotifiedhearings.PartiesNotifiedResponses;
+import uk.gov.hmcts.reform.hmc.model.unnotifiedhearings.UnNotifiedHearingResponse;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
